requirements.txt
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0
pyarrow>=12.0.0
tenacity>=8.2.0
pytest>=7.4.0
pytest-cov>=4.1.0
ruff>=0.0.287
black>=23.7.0